+++ CHANGELOG.md
@@ -6,11 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased] - 2025-10-26
+## [2.3.0] – 2025-11-10
 ### Changed
 - Invalid structure error message:
   - Shows no details. When pressing 'Show More', the user will see why it's invalid.
-  - Second message shows all errors instead of the first detected one.
+  - Second message shows all errors instead of only the first detected one.
 
 ### Added
 - Structure validation: